Fix number overflow issues with adjust operations on date (#730)
* Fix number overflow issues with adjust operations on date

* Apply suggestions from code review


---------

+++ legend-pure-m4/src/main/java/org/finos/legend/pure/m4/coreinstance/primitive/date/AbstractDateWithSecond.java
@@ -37,7 +37,7 @@ abstract class AbstractDateWithSecond extends AbstractDateWithMinute
     }
 
     @Override
-    public PureDate addSeconds(int seconds)
+    public PureDate addSeconds(long seconds)
     {
         if (seconds == 0)
         {
